Update Google maps api version to 3.2
Upon adding a marker to the map, the new marker would not save.
The latitude and longitude was no longer being stored in the marker.
Therefore it was necessary to update to Google's newest getPosition
function. Also locked the version in the google maps url to
prevent the code from breaking in the future.